feat: added breeder endpoint to get cages
